# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D Himmatlal and Company API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /testimonials:
    get:
      operationId: getTestimonials
      tags: [testimonials]
      summary: List testimonials
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    put:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        shortDescription: { type: ["string", "null"] }
        fullDescription: { type: ["string", "null"] }
        price: { type: ["string", "null"] }
        contactForPrice: { type: boolean }
        imageUrl: { type: ["string", "null"] }
        galleryImages: { type: ["string", "null"] }
        fabric: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, slug, categoryId, contactForPrice, inStock, isFeatured, isActive, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        categoryId: { type: integer }
        shortDescription: { type: ["string", "null"] }
        fullDescription: { type: ["string", "null"] }
        price: { type: ["string", "null"] }
        contactForPrice: { type: boolean }
        imageUrl: { type: ["string", "null"] }
        galleryImages: { type: ["string", "null"] }
        fabric: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        isActive: { type: boolean }
      required: [name, slug, categoryId, contactForPrice, inStock, isFeatured, isActive]
...
    Testimonial:
      type: object
      properties:
        id: { type: integer }
        customerName: { type: string }
        review: { type: string }
        rating: { type: integer }
        imageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, customerName, review, rating, isActive, createdAt, updatedAt]
    TestimonialInput:
      type: object
      properties:
        customerName: { type: string }
        review: { type: string }
        rating: { type: integer }
        imageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
      required: [customerName, review, rating, isActive]
...