# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * D Himmatlal and Company API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  categoryId: number;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  /** @nullable */
  price?: string | null;
  contactForPrice: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  /** @nullable */
  fabric?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  isActive: boolean;
}