# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * D Himmatlal and Company API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  customerName: string;
  review: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}